tests/wpt/tests/css/css-shapes/shape-outside/values/shape-outside-ellipse-000.html

<!DOCTYPE html>
<html>
    <head>
        <title>Shape Outside Ellipse Valid Formats</title>
        <link rel="author" title="Adobe" href="http://html.adobe.com/">
        <link rel="author" title="Bear Travis" href="mailto:betravis@adobe.com">
        <link rel="reviewer" title="Alan Stearns" href="mailto:stearns@adobe.com"> <!-- 2014-03-04 -->
        <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#funcdef-ellipse">
        <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#shape-outside-property">
        <meta name="assert" content="An elliptical basic shape has two optional components, radii (2) and a position.">
        <meta name="flags" content="ahem dom">
        <meta name="viewport" content="width=device-width,initial-scale=1,minimum-scale=1">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="support/parsing-utils.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script type="text/javascript">
            var valid_ellipse_tests = [
                {
                  "name": "No arguments",
                  "actual": "ellipse()",
                  "expected_inline": "ellipse()",
                  "expected_computed": "ellipse()"
                },
                {
                  "name": "Radial arguments",
                  "actual": "ellipse(50px 50px)",
                  "expected_inline": "ellipse(50px 50px)",
                  "expected_computed": "ellipse(50px 50px)"
                },
                {
                   "name": "Position argument",
                   "actual": "ellipse(at 50px 50px)",
                   "expected_inline": "ellipse(at 50px 50px)",
                   "expected_computed": "ellipse(at 50px 50px)"
                },
                {
                   "name": "Radial and position argument",
                   "actual": "ellipse(50px 50px at 50px 50px)",
                   "expected_inline": "ellipse(50px 50px at 50px 50px)",
                   "expected_computed": "ellipse(50px 50px at 50px 50px)"
                }
            ];
            generate_tests( ParsingUtils.testInlineStyle,
                            ParsingUtils.buildTestCases(valid_ellipse_tests, "inline") );
            generate_tests( ParsingUtils.testComputedStyle,
                            ParsingUtils.buildTestCases(valid_ellipse_tests, "computed") );
        </script>
    </body>
</html>
