third_party/blink/web_tests/external/wpt/css/css-shapes/shape-outside/values/shape-outside-ellipse-000-expected.txt

This is a testharness.js-based test.
FAIL No arguments - inline assert_equals: expected "ellipse(at 50% 50%)" but got "ellipse()"
FAIL Radial arguments - inline assert_equals: expected "ellipse(50px 50px at 50% 50%)" but got "ellipse(50px 50px)"
PASS Position argument - inline
PASS Radial and position argument - inline
FAIL No arguments - computed assert_equals: expected "ellipse(at 50% 50%)" but got "ellipse()"
FAIL Radial arguments - computed assert_equals: expected "ellipse(50px 50px at 50% 50%)" but got "ellipse(50px 50px)"
PASS Position argument - computed
PASS Radial and position argument - computed
Harness: the test ran to completion.
